consensus: extend block timestamps past 2106
Consensus changes in this draft:

- The first block with nTime >= 2^32 must have a negative nVersion and use the new 84-byte header format.

- Every subsequent block must continue to use the new header format.

- Blocks using the new header format must have nTime >= 2^32.

- Block timestamps, MTP, and timestamp comparisons used by consensus validation are represented as uint64_t, allowing the full uint64_t timestamp range.

Tests added/updated:

- Add feature_y2106_extended_time.py to exercise the post-2106 hardfork behavior against current and previous-release nodes, including the max uint64 timestamp reaching contextual validation.

- Extend interface_ipc_mining.py to cover 64-bit submitSolution timestamps.

+++ src/interfaces/mining.h
@@ -75,7 +75,7 @@ public:
      *       the solved block is constructed and broadcast by multiple nodes
      *       (e.g. both the miner who constructed the template and the pool).
      */
-    virtual bool submitSolution(uint32_t version, uint32_t timestamp, uint32_t nonce, CTransactionRef coinbase) = 0;
+    virtual bool submitSolution(uint32_t version, uint64_t timestamp, uint32_t nonce, CTransactionRef coinbase) = 0;
 
     /**
      * Waits for fees in the next block to rise, a new tip or the timeout.
